# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Authentication and User Management API",
    "version": "1.0.0",
    "description": "This API provides endpoints for managing user authentication, authorization, and profile operations. It includes features like user registration, login, two-factor authentication, password and email updates, and various system-level functionalities such as retrieving hooks and states. Designed to support secure and scalable authentication workflows."
  },
  "servers": [
    {
      "url": "http://localhost:80"
    }
  ],
...
  "paths": {
...
    "/api/user/{userId}": {
      "get": {
        "tags": ["User"],
        "summary": "Get user details",
        "description": "Retrieves details of a user by their ID. Only admins or the user themselves can access this endpoint. The Bearer token required for authorization can be obtained by calling `/api/authentication/signin`.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user to retrieve."
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Bearer token for authentication, obtained from `/api/authentication/signin`."
          },
          {
            "name": "x-xsrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "XSRF token obtained from the `/api/system/xsfr` endpoint"
          }
        ],
        "responses": {
          "200": {
            "description": "User details retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "description": "Unique identifier of the user.",
                      "example": "5f8d0d55b54764421b7156fe"
                    },
                    "ip": {
                      "type": "string",
                      "description": "The IP address of the user.",
                      "example": "192.168.1.1"
                    },
                    "username": {
                      "type": "string",
                      "description": "The username of the user.",
                      "example": "johndoe"
                    },
                    "email": {
                      "type": "string",
                      "format": "email",
                      "description": "The email address of the user.",
                      "example": "johndoe@example.com"
                    },
                    "role": {
                      "type": "string",
                      "description": "The role of the user (e.g., ADMIN, USER).",
                      "example": "user"
                    },
                    "state": {
                      "type": "string",
                      "description": "The current state of the user. (.e.g. `ACTIVE`, `PENDING`, `BLOCKED`)",
                      "example": "ACTIVE"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "The timestamp when the user was created.",
                      "example": "2023-01-01T12:00:00Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "The timestamp when the user was last updated.",
                      "example": "2023-06-01T12:00:00Z"
                    },
                    "lastLogin": {
                      "type": "string",
                      "format": "date-time",
                      "description": "The timestamp of the user's last login.",
                      "example": "2023-07-01T10:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing parameter `userId`"
          },
          "403": {
            "description": "The user is not authorized to access this resource."
          },
          "404": {
            "description": "User not found."
          },
          "406": {
            "description": "User is not ACTIVE."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/api/user/modify/email": {
      "patch": {
        "tags": ["User"],
        "summary": "Modify user email",
        "description": "Creates an email modification object that allows the user to change their email address. An email is sent to confirm the change. The Bearer token required for authorization can be obtained by calling `/api/authentication/signin`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The new email address to set.",
                    "example": "newemail@example.com"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Bearer token for authentication, obtained from `/api/authentication/signin`."
          },
          {
            "name": "x-xsrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "XSRF token obtained from the `/api/system/xsfr` endpoint"
          }
        ],
        "responses": {
          "200": {
            "description": "Email modification object created successfully. A confirmation email has been sent."
          },
          "400": {
            "description": "Missing / invalid email or invalid authentication token."
          },
          "403": {
            "description": "The user is not authorized to access this resource."
          },
          "404": {
            "description": "User not found."
          },
          "409": {
            "description": "Conflict. Email address already exists."
          },
          "406": {
            "description": "User is not ACTIVE."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/api/user/modify/password": {
      "patch": {
        "tags": ["User"],
        "summary": "Modify user password",
        "description": "Creates a password modification object that allows the user to change their password. An email is sent to confirm the change. The Bearer token required for authorization can be obtained by calling `/api/authentication/signin`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string",
                    "description": "The current password of the user.",
                    "example": "oldpassword123"
                  },
                  "newPassword": {
                    "type": "string",
                    "description": "The new password to set.",
                    "example": "newpassword456"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Bearer token for authentication, obtained from `/api/authentication/signin`."
          },
          {
            "name": "x-xsrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "XSRF token obtained from the `/api/system/xsfr` endpoint"
          }
        ],
        "responses": {
          "200": {
            "description": "Password modification object created successfully. A confirmation email has been sent."
          },
          "400": {
            "description": "Missing or invalid currentPassword or newPassword or invalid authentication token."
          },
          "403": {
            "description": "The user is not authorized to access this resource or the current password is incorrect."
          },
          "404": {
            "description": "User not found."
          },
          "406": {
            "description": "User is not ACTIVE."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    }
  }
}